Second commit, more ts conversions
