+++ src/sequencer/grid/animation/interfaces/AnimationContext.ts
@@ -1,0 +1,4 @@
+// Define the AnimationContext type
+export interface AnimationContext {
+    animationCtx: CanvasRenderingContext2D;
+}
