--- src/sequencer/grid/drawing/color-schemes/grid-colors.js
@@ -1,60 +1,0 @@
-// src/sequencer/grid/drawing/color-schemes/grid-colors.js
-
-export const GRID_COLOR_SCHEMES = {
-    Midnight: {
-      whiteKey: '#1e1e1e',
-      blackKey: '#2a103b',
-      labelWhite: '#2c2c2c',
-      labelBlack: '#3a1d4d',
-      textWhite: '#cfcfcf',
-      textBlack: '#e6d9ff',
-      gridLine: '#333',
-      beatLine: '#4e3d63',
-      measureLine: '#8561c2'
-    },      
-    Seashell: {
-      whiteKey: '#fefefe',
-      blackKey: '#f1e7dc',
-      labelWhite: '#dddddd',
-      labelBlack: '#d4cfc9',
-      textWhite: '#444',
-      textBlack: '#663300',
-      gridLine: '#ccc',
-      beatLine: '#bbb',
-      measureLine: '#999'
-    },
-    Cyberpunk: {
-      whiteKey: '#1b1b2f',
-      blackKey: '#162447',
-      labelWhite: '#1f4068',
-      labelBlack: '#1f4068',
-      textWhite: '#e43f5a',
-      textBlack: '#e43f5a',
-      gridLine: '#1f4068',
-      beatLine: '#e43f5a',
-      measureLine: '#ffcc00'
-    },
-    'Classic Blue': {
-      whiteKey: '#f5faff',
-      blackKey: '#dceeff',
-      labelWhite: '#aaccee',
-      labelBlack: '#88bbdd',
-      textWhite: '#002244',
-      textBlack: '#001122',
-      gridLine: '#99bbdd',
-      beatLine: '#88aacc',
-      measureLine: '#336699'
-    },
-    Darkroom: {
-      whiteKey: '#2c2c2c',
-      blackKey: '#1e1e1e',
-      labelWhite: '#383838',
-      labelBlack: '#2a2a2a',
-      textWhite: '#a0a0a0',
-      textBlack: '#e0e0e0',
-      gridLine: '#404040',
-      beatLine: '#606060',
-      measureLine: '#808080'
-    }
-  };
-
