+++ src/sequencer/grid/drawing/rounded-rect.ts
@@ -1,0 +1,32 @@
+// src/sequencer/grid/drawing/rounded-rect.ts
+
+/**
+ * Draws a rounded rectangle path on the canvas.
+ *
+ * @param ctx - Canvas rendering context
+ * @param x - X position
+ * @param y - Y position
+ * @param w - Width
+ * @param h - Height
+ * @param r - Corner radius (default = 4)
+ */
+export function drawRoundedRect(
+  ctx: CanvasRenderingContext2D,
+  x: number,
+  y: number,
+  w: number,
+  h: number,
+  r: number = 4
+): void {
+  ctx.beginPath();
+  ctx.moveTo(x + r, y);
+  ctx.lineTo(x + w - r, y);
+  ctx.quadraticCurveTo(x + w, y, x + w, y + r);
+  ctx.lineTo(x + w, y + h - r);
+  ctx.quadraticCurveTo(x + w, y + h, x + w - r, y + h);
+  ctx.lineTo(x + r, y + h);
+  ctx.quadraticCurveTo(x, y + h, x, y + h - r);
+  ctx.lineTo(x, y + r);
+  ctx.quadraticCurveTo(x, y, x + r, y);
+  ctx.closePath();
+}
