+++ src/sequencer/grid/helpers/constants.ts
@@ -1,6 +1,8 @@
-export const labelWidth = 64;
+// src/sequencer/grid/helpers/constants.ts
 
-export const PITCH_COLOR_MAP = {
+export const labelWidth = 64 as const;
+
+export const PITCH_COLOR_MAP: Record<string, string> = {
   'C': '#FF0000',
   'C#': '#9400D3',
   'D': '#FFFF00',
@@ -19,7 +21,12 @@ export const HEIGHT_RATIO = 0.15;
 export const MIN_CELL_WIDTH = 100;
 export const MAX_CELL_WIDTH = 300;
 
-export const ZOOM_LEVELS = [
+export interface ZoomLevel {
+  cellWidth: number;
+  cellHeight: number;
+}
+
+export const ZOOM_LEVELS: ZoomLevel[] = [
   { cellWidth: 20, cellHeight: 10 },
   { cellWidth: 30, cellHeight: 15 },
   { cellWidth: 40, cellHeight: 20 }, // ⬅️ default
